skillsync-ml/requirements.txt
sentence-transformers==2.2.2
torch==2.7.0
transformers==4.26.1
huggingface_hub==0.13.3
grpcio==1.47.0
grpcio-tools==1.47.0
numpy==1.24.2
scipy==1.10.1
protobuf==3.20.3
